remove references to old identity newsletter models
