+++ common/app/services/newsletters/model/EmailEmbed.scala
@@ -1,0 +1,11 @@
+package services.newsletters.model
+
+case class EmailEmbed(
+    name: String,
+    title: String,
+    description: String,
+    successHeadline: String,
+    successDescription: String,
+    hexCode: String,
+    imageUrl: Option[String],
+)
