+++ common/app/services/newsletters/model/NewsletterIllustration.scala
@@ -1,0 +1,5 @@
+package services.newsletters.model
+
+case class NewsletterIllustration(
+    circle: Option[String],
+)
