+++ common/app/views/fragments/email/signup/result/emailSuccess.scala.html
@@ -1,4 +1,4 @@
-@import com.gu.identity.model.EmailEmbed
+@import services.newsletters.model.EmailEmbed
 @(emailEmbedData: EmailEmbed)
 <div class="email-sub email-sub--article">
     <div class="email-sub__message">
